+++ CHANGELOG.md
@@ -54,6 +54,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - (**Migration**) Fix selecting a destination from a source browse search page
 - (**Migration**) Fix being able to migrate an entry to itself
 - (**Migration**) Fix resuming migration when app is opened in multiple tabs
+- (**Migration**) Fix showing "All source searches failed" error while the entry has a selected destination source or matches were found
 - (**Reader**) Fix scrollbar appearing with "fit to widt/height/screen" page scale mode and applied safe area insets
 - (**Reader**) Fix wrongly positioned mobile progress bar current page indicator
 - (**Reader**) Fix mobile progress bar previous/next chapter button visibility on hover and while disabled